+++ src/components/Profile/ProfileSection/styledProfileSection.jsx
@@ -1,0 +1,33 @@
+import styled from "styled-components";
+
+const Section = styled.section`
+  width: 100%;
+  padding: 30px 24px;
+  background: white;
+  text-align: center;
+
+  strong {
+    display: block;
+    font-size: 16px;
+    font-weight: 700;
+    margin: 10px 0;
+  }
+
+  small {
+    display: block;
+    color: #767676;
+    font-size: 12px;
+    font-weight: 400;
+    line-height: 14px;
+    margin: 10px 0;
+  }
+
+  p {
+    display: block;
+    line-height: 17.53px;
+    font-weight: 400;
+    margin: 10px 0;
+  }
+`;
+
+export { Section };
